Add files via upload
In this experiment, we introduce the Variational Information Bottleneck (VIB) objective function to train the weight imprinting model.
These works merge the imprinted-weight model proposed by Qi et al. and VIB loss.
Most of the are credited to https://github.com/YU1ut

We used a pre-trained model resnet50.
Download the CUB-200-2011 Dataset. 
Save in a file called "Data". 
The file run.sh and pre-train contain hyperparameters that can be tuned such as latent dimension, beta, etc.

Execute the run.sh file
The results of the pre-trained model will be stored in "save_pretained" and the imprinting will be saved in "save_imprinting".